+++ src/lib.rs
@@ -15,9 +15,10 @@ mod server;
 
 use antidote::Mutex;
 use std::collections::HashMap;
+use std::vec::Vec;
 
 lazy_static! {
-    pub(crate) static ref MAP: Mutex<HashMap<u8, Request>> = Mutex::new(HashMap::new());
+    pub(crate) static ref QUEUE: Mutex<HashMap<u8, Vec<Request>>> = Mutex::new(HashMap::new());
 }
 
 #[derive(Debug)]
@@ -28,6 +29,6 @@ pub struct Request {
     pub path: String,
 }
 
-pub(crate) struct SendRequest {
-    tx: channel::Sender<u8>,
+pub(crate) struct ShareRequest {
+    id: u8,
 }